DigitalNote v1.0.1.5 Update 1
Changelog

- Fixed rpcmining getblocktemplate
- General cleanup
- Added MasterNode checks delay system to later engage in-depth per-block security without affecting sync or operational stability
- Loosened up timeout protocol to keep connection counts higher with the network (previous were unrealisticly stringent)
- Added ping timeout parameters

+++ src/rpcmining.cpp
@@ -687,7 +687,7 @@ Value getblocktemplate(const Array& params, bool fHelp)
             CAmount devopsSplit = devopsPayment;
             Object devopsReward;
             devopsReward.push_back(Pair("devopspayee", Params().DevOpsAddress()));
-            devopsReward.push_back(Pair("amount", devopsSplit));
+            devopsReward.push_back(Pair("amount", (int64_t)devopsSplit));
             result.push_back(Pair("devopsreward", devopsReward));
             result.push_back(Pair("devops_reward_enforced", true));
